fix configuration on OG pebbles (by reducing memory usage?); only update temp expiration if current temp is available

+++ src/c/ticks.h
@@ -4,4 +4,4 @@
 void init_ticks(Layer *layer);
 void update_battery_ticks(BatteryChargeState charge_state);
 void animate_charging_indicator();
-void update_tick_settings(GColor colour, int size, bool battery_enabled);
+void update_tick_settings(GColor colour, uint8_t size, bool battery_enabled);
